Move all the extensions interface under a same directory (#650)
