+++ src/main/java/org/opensearch/sdk/api/MapperExtension.java
@@ -7,10 +7,11 @@
  * compatible open source license.
  */
 
-package org.opensearch.sdk;
+package org.opensearch.sdk.api;
 
 import org.opensearch.index.mapper.Mapper;
 import org.opensearch.index.mapper.MetadataFieldMapper;
+import org.opensearch.sdk.Extension;
 
 import java.util.Collections;
 import java.util.Map;
