+++ src/main/java/org/opensearch/sdk/api/ReloadableExtension.java
@@ -7,9 +7,10 @@
  * compatible open source license.
  */
 
-package org.opensearch.sdk;
+package org.opensearch.sdk.api;
 
 import org.opensearch.common.settings.Settings;
+import org.opensearch.sdk.Extension;
 
 /**
  * An extension point for {@link Extension}s that can be reloaded. There is no
