+++ src/main/java/org/opensearch/sdk/api/SystemIndexExtension.java
@@ -7,7 +7,7 @@
  * compatible open source license.
  */
 
-package org.opensearch.sdk;
+package org.opensearch.sdk.api;
 
 import org.opensearch.common.settings.Settings;
 import org.opensearch.indices.SystemIndexDescriptor;
