+++ src/test/java/org/opensearch/sdk/action/TestSDKActionModule.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.Test;
 import org.opensearch.action.ActionRequest;
 import org.opensearch.action.ActionResponse;
 import org.opensearch.action.ActionType;
-import org.opensearch.sdk.ActionExtension;
+import org.opensearch.sdk.api.ActionExtension;
 import org.opensearch.sdk.BaseExtension;
 import org.opensearch.sdk.ExtensionSettings;
 import org.opensearch.test.OpenSearchTestCase;
